requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
lxml>=4.9.0
